fix: archunit-module-scope
